Insurance Broker Home UI
For the title we have given it a slot start and padding at the start.

The logo is fixed at the center and given transform property to maintain it's position on any other device.

This makes it fully responsive.

The ion-fab-button holds the avatar and a nice blue border.

The other buttons have secondary colors to represent the brand.